+++ lib/src/main/java/formatter/SequenceParserResult.java
@@ -4,6 +4,6 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class SequenceParserResult {
-    public float[] sequence;
+    public double[] sequence;
     public List<String> errors = new ArrayList<>();
 }
